scorecanvas
move all the score graphics code to scorecanvas.js
