+++ src/Game.js
@@ -41,6 +41,13 @@ import {
   initPickUI,
   selectDirection
 } from "./graphics.js";
+
+import {
+  scoreCanvasInit,
+  clearScore,
+  updateScores
+} from "./scorecanvas.js";
+
 /**
  * Created with JetBrains WebStorm.
  * User: dmaynard
@@ -54,7 +61,7 @@ import {
 var nextToMove;
 var focusPoint;
 var focusWorm
-var scorectx;
+
 // the jump from full pan left (-1.0) to full pan right (+1.0)
 // is too jaring. This limits pan to [-.8 , +.8]
 // Could be a setting  (pan effect 0 - 1)
@@ -332,35 +339,6 @@ export function binToRGB(bin) {
 
 //  Called from Timer Loop
 
-function clearScore(segmentIndex, totalSegments) {
-  var segWidth = darworms.dwsettings.scoreCanvas.width / totalSegments;
-  scorectx.fillStyle = "rgba(222,222,222, 1.0)";
-  scorectx.shadowOffsetX = 0;
-  scorectx.shadowOffsetY = 0;
-
-  scorectx.fillRect(segWidth * segmentIndex, 0, segWidth, darworms.dwsettings.scoreCanvas.height);
-}
-
-function scoreStartx(segmentIndex, totalSegments, text) {
-  var segWidth = darworms.dwsettings.scoreCanvas.width / totalSegments;
-  var twidth = scorectx.measureText(text).width;
-  return ((segWidth * segmentIndex) + (segWidth / 2) - (twidth / 2));
-
-}
-
-function updateScores() {
-  var i;
-  for (i = 0; i < 4; i++) {
-    if (darworms.theGame.worms[i] !== undefined && darworms.theGame.worms[i].shouldDrawScore()) {
-      clearScore(i, 4);
-      scorectx.fillStyle = darworms.dwsettings.colorTable[i + 1];
-      // scorectx.shadowOffsetX = 3;
-      // scorectx.shadowOffsetY = 3;
-      scorectx.fillText(darworms.theGame.worms[i].score, scoreStartx(i, 4, darworms.theGame.worms[i].score.toString()), 15);
-    }
-  }
-};
-
 function makeMoves() {
   // console.log(" makeMoves theGameOver " + theGameOver +  "  gameState " + gameStateNames[theGame.gameState] );
   var startTime = new Date().getTime();
@@ -398,15 +376,8 @@ function gameInit() {
   // used to initialize variables in this module's closure
   console.log(" wCanvas,width: " + wCanvas.width);
   graphicsInit(this);
+  scoreCanvasInit();
   nextToMove = 0;
-  window.scoreCanvas = document.getElementById("scorecanvas");
-  scorectx = darworms.dwsettings.scoreCanvas.getContext("2d");
-  scorectx.font = "bold 18px sans-serif";
-  scorectx.shadowColor = "rgb(190, 190, 190)";
-  scorectx.shadowOffsetX = 3;
-  scorectx.shadowOffsetY = 3;
-
-
 
 }
